sistema_patrimonio_entregavel/requirements.txt
flask
openpyxl
gunicorn

google-api-python-client
google-auth-httplib2
google-auth-oauthlib
